+++ src/schemas/question.ts
@@ -59,6 +59,8 @@ type Question {
 
     "The conditional logic triggered by this question"
     questionConditions: [QuestionCondition!]
+    "The question options associated with this question"
+    questionOptions: [QuestionOption!]
 }
 
 input AddQuestionInput {
@@ -99,5 +101,31 @@ input UpdateQuestionInput {
     sampleText: String
     "To indicate whether the question is required to be completed"
     required: Boolean
+}
+
+"QuestionOption always belongs to a Question"
+type QuestionOption {
+    "The unique identifer for the Object"
+    id: Int
+    "The user who created the Object"
+    createdById: Int
+    "The timestamp when the Object was created"
+    created: String
+    "The user who last modified the Object"
+    modifiedById: Int
+    "The timestamp when the Object was last modifed"
+    modified: String
+    "Errors associated with the Object"
+    errors: [String!]
+
+    "The question id that the QuestionOption belongs to"
+    questionId: Int!
+    "The option text"
+    text: String!
+    "The option order number"
+    orderNumber: Int!
+    "Whether the option is the default selected one"
+    isDefault: Boolean
+
 }
 `